requirement.txt
Django>=4.0,<5.0
djangorestframework
requests
APScheduler
drf-yasg
python-dotenv
